ValueBuildersProject/vb_pro/requirements.txt
aiofiles==23.2.1
anyio==4.2.0
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
frontend==0.0.3
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
joblib==1.3.2
lxml==5.1.0
nltk==3.8.1
pillow==10.2.0
pytils==0.4.1
pytz==2023.4
regex==2023.12.25
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.36.1
tools==0.1.9
tqdm==4.66.1
tzdata==2023.4
urllib3==2.2.0
uvicorn==0.27.0.post1
